# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SwiftEntry API",
        "description": "Backend REST de SwiftEntry \u2014 plataforma de venta de boletos para eventos.\n\n**Autenticaci\u00f3n:** todas las rutas protegidas requieren un JWT en el header `Authorization: Bearer <token>`. Obt\u00e9n el token en `POST /swift_entry/auth/login` y p\u00e9galo en el bot\u00f3n **Authorize** de arriba.\n\n**Flujo principal:** Evento \u2192 Localidad \u2192 Asiento \u2192 Reserva \u2192 Pago \u2192 Boleto\n",
        "contact": {
            "name": "SwiftEntry Team",
            "email": "gerardo.port1107.psql@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
        "/swift_entry/localities/{id}": {
            "get": {
                "tags": [
                    "Localidades"
                ],
                "summary": "Obtener localidad por ID",
                "operationId": "getLocalityById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Localidades"
                ],
                "summary": "Actualizar localidad",
                "operationId": "updateLocality",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LocalityUpdateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Localidades"
                ],
                "summary": "Eliminar localidad",
                "operationId": "deleteLocality",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/swift_entry/payments": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Procesar pago",
                "description": "Convierte una reserva PENDING en CONFIRMED, marca los asientos como SOLD y emite un Ticket por cada asiento. Incluye IVA del 13%",
                "operationId": "processPayment",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/swift_entry/localities": {
            "get": {
                "tags": [
                    "Localidades"
                ],
                "summary": "Listar todas las localidades",
                "description": "Ruta p\u00fablica",
                "operationId": "getAllLocalities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Localidades"
                ],
                "summary": "Crear localidad",
                "description": "\u26a0\ufe0f Stub \u2014 actualmente no hace nada. Las localidades se crean desde el endpoint de Eventos",
                "operationId": "createLocality",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LocalityRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/swift_entry/seats/locality/{localityId}": {
            "get": {
                "tags": [
                    "Asientos"
                ],
                "summary": "Asientos de una localidad",
                "operationId": "getSeatsByLocalityId",
                "parameters": [
                    {
                        "name": "localityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/swift_entry/localities/event/{eventId}": {
            "get": {
                "tags": [
                    "Localidades"
                ],
                "summary": "Localidades de un evento",
                "operationId": "getLocalitiesByEventId",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "LocalityUpdateDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "minimum": 0.0
                    },
                    "earlyBirdDiscountPercentage": {
                        "type": "number",
                        "maximum": 100.0,
                        "minimum": 0.0
                    },
                    "earlyBirdDeadline": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "capacity": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "availableSlots": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    }
                }
            },
            "GeneralResponse": {
                "type": "object",
                "properties": {
                    "uri": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {}
                }
            },
...
            "PaymentRequestDTO": {
                "type": "object",
                "properties": {
                    "reservationId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "CASH",
                            "CREDIT_CARD",
                            "DEBIT_CARD",
                            "BANK_TRANSFER"
                        ]
                    }
                },
                "required": [
                    "paymentMethod",
                    "reservationId"
                ]
            },
            "LocalityRequestDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "minimum": 0.0
                    },
                    "earlyBirdDiscountPercentage": {
                        "type": "number",
                        "maximum": 100.0,
                        "minimum": 0.0
                    },
                    "earlyBirdDeadline": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "name",
                    "price"
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}